(refact) Move ratingInfo file back out of models folder.
